+++ CHANGELOG.md
@@ -5,6 +5,8 @@ Common and useful classes, methods, exceptions etc.
 # 1.0.3
 
 1. Travis CI > run many tasks using Phing > update configuration
+2. Template with placeholders > verification of placeholders without values > make stronger and point out which are 
+missing
 
 # 1.0.2